requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.9
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   langchain-core
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
dataclasses-json==0.6.7
    # via
    #   langchain
    #   langchain-community
fastapi==0.104.1
    # via -r requirements.in
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.4.0
    # via google-generativeai
google-api-core[grpc]==2.25.0
    # via
    #   google-ai-generativelanguage
    #   google-generativeai
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-generativeai
google-generativeai==0.3.2
    # via langchain-google-genai
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.3
    # via sqlalchemy
grpcio==1.72.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
joblib==1.5.1
    # via
    #   nltk
    #   scikit-learn
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.1.0
    # via -r requirements.in
langchain-community==0.0.20
    # via langchain
langchain-core==0.1.23
    # via
    #   langchain
    #   langchain-community
    #   langchain-google-genai
langchain-google-genai==0.0.6
    # via -r requirements.in
langsmith==0.0.87
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
nltk==3.8.1
    # via -r requirements.in
numpy==1.24.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   pgvector
    #   scikit-learn
    #   scipy
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pgvector==0.2.4
    # via -r requirements.in
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.8
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.5.0
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.14.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.6
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
regex==2024.11.6
    # via nltk
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
rsa==4.9.1
    # via google-auth
scikit-learn==1.3.2
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.23
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
starlette==0.27.0
    # via fastapi
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   google-generativeai
    #   nltk
typing-extensions==4.14.0
    # via
    #   fastapi
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.4.0
    # via requests
uvicorn[standard]==0.24.0
    # via -r requirements.in
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.0
    # via aiohttp
